requirements.txt
streamlit>=1.29
python-docx>=1.1.0
pypdf>=4.0.0
faster-whisper>=1.0.3
pytest>=7.4
